Put scatter and features scatter tests in different files
