+++ src/napari_matplotlib/tests/scatter/baseline/test_features_scatter_widget_2D.png
