+++ src/napari_matplotlib/tests/scatter/baseline/test_scatter_2D.png
